client/requirements.txt
pyautogui
pillow
requests
python-dotenv
tk
pyttsx3
pynput
google-cloud-texttospeech
google-cloud-speech
PyAudio
